+++ src/commands/configure.integration.test.ts
@@ -116,13 +116,17 @@ describe('Configure CLI command', () => {
         '🆗 All security schemes have been configured successfully.'
       );
       expect(result.stdout).toMatch(
-        '🆗 Parameter version configured with default value "v1"'
+        `Provider azure-cognitive-services has integration parameters that must be configured. You can configure them in super.json on path: superface/super.json or set the environment variables as defined below.`
+      );
+      expect(result.stdout).toMatch(
+        '🆗 Parameter version has been configured to use value of environment value "$AZURE_COGNITIVE_SERVICES_VERSION"'
       );
       expect(result.stdout).toContain(
-        '❌ Parameter instance with description "Instance of your azure cognitive service" has not been configured.'
+        'Please, configure this environment value.'
       );
+
       expect(result.stdout).toContain(
-        'Please, configure this parameter manualy in super.json on path: superface/super.json'
+        'If you do not set the variable, the default value "v1" will be used.'
       );
 
       await expect(
@@ -146,7 +150,7 @@ describe('Configure CLI command', () => {
         superJson.normalized.providers[providerWithParameters].parameters
       ).toEqual({
         instance: '$AZURE_COGNITIVE_SERVICES_INSTANCE',
-        version: 'v1',
+        version: '$AZURE_COGNITIVE_SERVICES_VERSION',
       });
       expect(superJson.document.profiles![profileId]).toEqual({
         version: profileVersion,