+++ src/logic/check.test.ts
@@ -1,5 +1,6 @@
 import {
   ApiKeyPlacement,
+  AstMetadata,
   HttpScheme,
   MapDocumentNode,
   ProfileDocumentNode,
@@ -18,6 +19,7 @@ import {
 import { ProfileId } from '../common/profile';
 import {
   check,
+  checkIntegrationParameters,
   checkMapAndProfile,
   checkMapAndProvider,
   CheckResult,
@@ -70,8 +72,23 @@ describe('Check logic', () => {
   const mockProfileSource = 'mock profile source';
   const mockMapSource = 'mock map source';
 
+  const astMetadata: AstMetadata = {
+    sourceChecksum: 'check',
+    astVersion: {
+      major: 1,
+      minor: 0,
+      patch: 0,
+    },
+    parserVersion: {
+      major: 1,
+      minor: 0,
+      patch: 0,
+    },
+  };
+
   const mockMapDocument: MapDocumentNode = {
     kind: 'MapDocument',
+    astMetadata,
     header: {
       kind: 'MapHeader',
       profile: {
@@ -97,6 +114,7 @@ describe('Check logic', () => {
 
   const mockMapDocumentWithUnverified: MapDocumentNode = {
     kind: 'MapDocument',
+    astMetadata,
     header: {
       kind: 'MapHeader',
       profile: {
@@ -122,6 +140,7 @@ describe('Check logic', () => {
 
   const mockProfileDocument: ProfileDocumentNode = {
     kind: 'ProfileDocument',
+    astMetadata,
     header: {
       kind: 'ProfileHeader',
       name: 'character-information',
@@ -232,7 +251,7 @@ describe('Check logic', () => {
   };
 
   describe('when checking capability', () => {
-    it('returns correctly formated string when we use local files', async () => {
+    it('returns correctly formated result when we use local files', async () => {
       const mockProfile = {
         name: 'character-information',
       };
@@ -290,6 +309,13 @@ describe('Check logic', () => {
           profileId: 'starwars/character-information',
           provider: unverifiedProvider,
         },
+        {
+          kind: 'parameters',
+          providerFrom: mockLocalProviderFrom,
+          issues: [],
+          provider: unverifiedProvider,
+          superJsonPath: '',
+        },
       ]);
 
       expect(loadProvider).toHaveBeenCalledWith(
@@ -316,7 +342,7 @@ describe('Check logic', () => {
       expect(fetchProviderInfo).not.toHaveBeenCalled();
     });
 
-    it('returns correctly formated string when we use remote files', async () => {
+    it('returns correctly formated result when we use remote files', async () => {
       const mockSuperJson = new SuperJson({
         profiles: {
           [`${profile.scope}/${profile.name}`]: {
@@ -371,6 +397,13 @@ describe('Check logic', () => {
           issues: [],
           profileId: 'starwars/character-information',
           provider: unverifiedProvider,
+        },
+        {
+          kind: 'parameters',
+          providerFrom: mockRemoteProviderFrom,
+          issues: [],
+          provider: unverifiedProvider,
+          superJsonPath: '',
         },
       ]);
 
@@ -709,6 +742,30 @@ describe('Check logic', () => {
             },
           ],
         },
+        {
+          kind: 'parameters',
+          provider,
+          providerFrom: mockLocalProviderFrom,
+          superJsonPath: 'some/path',
+          issues: [
+            {
+              kind: 'error',
+              message: 'third-check-first-error',
+            },
+            {
+              kind: 'warn',
+              message: 'third-check-first-warn',
+            },
+            {
+              kind: 'error',
+              message: 'third-check-second-error',
+            },
+            {
+              kind: 'warn',
+              message: 'third-check-second-warn',
+            },
+          ],
+        },
       ];
       const formated = formatHuman(mockResult);
       //First title
@@ -727,11 +784,23 @@ describe('Check logic', () => {
         `for profile ${profile.scope}/${profile.name} and local provider ${provider} at path`
       );
       expect(formated).toMatch(mockLocalProviderFrom.path);
-      //Second title
+      // Second body
       expect(formated).toMatch('❌ second-check-first-error');
       expect(formated).toMatch('⚠️ second-check-first-warn');
       expect(formated).toMatch('❌ second-check-second-error');
       expect(formated).toMatch('⚠️ second-check-second-warn');
+      //Third title
+      expect(formated).toMatch(
+        `❌ Checking integration parameters of local provider at path`
+      );
+      expect(formated).toMatch(mockLocalProviderFrom.path);
+      expect(formated).toMatch(`and super.json at path`);
+      expect(formated).toMatch('some/path');
+      //Third body
+      expect(formated).toMatch('❌ third-check-first-error');
+      expect(formated).toMatch('⚠️ third-check-first-warn');
+      expect(formated).toMatch('❌ third-check-second-error');
+      expect(formated).toMatch('⚠️ third-check-second-warn');
     });
   });
 
@@ -788,6 +857,7 @@ describe('Check logic', () => {
     it('returns result with errors if profile and map has different name,scope and version', async () => {
       const mockMapDocument: MapDocumentNode = {
         kind: 'MapDocument',
+        astMetadata,
         header: {
           kind: 'MapHeader',
           profile: {
@@ -807,6 +877,7 @@ describe('Check logic', () => {
 
       const mockProfileDocument: ProfileDocumentNode = {
         kind: 'ProfileDocument',
+        astMetadata,
         header: {
           kind: 'ProfileHeader',
           name: 'character-information',
@@ -828,6 +899,7 @@ describe('Check logic', () => {
     it('returns result with errors and warnings if profile and map has different name,scope, version and number of usecasses', async () => {
       const mockMapDocument: MapDocumentNode = {
         kind: 'MapDocument',
+        astMetadata,
         header: {
           kind: 'MapHeader',
           profile: {
@@ -847,6 +919,7 @@ describe('Check logic', () => {
 
       const mockProfileDocument: ProfileDocumentNode = {
         kind: 'ProfileDocument',
+        astMetadata,
         header: {
           kind: 'ProfileHeader',
           name: 'character-information',
@@ -879,6 +952,7 @@ describe('Check logic', () => {
     it('returns result with errors if profile and map has different name,scope, version and number of usecasses -strict', async () => {
       const mockMapDocument: MapDocumentNode = {
         kind: 'MapDocument',
+        astMetadata,
         header: {
           kind: 'MapHeader',
           profile: {
@@ -898,6 +972,7 @@ describe('Check logic', () => {
 
       const mockProfileDocument: ProfileDocumentNode = {
         kind: 'ProfileDocument',
+        astMetadata,
         header: {
           kind: 'ProfileHeader',
           name: 'character-information',
@@ -948,6 +1023,7 @@ describe('Check logic', () => {
     it('returns result with errors if provider and map has different name', async () => {
       const mockMapDocument: MapDocumentNode = {
         kind: 'MapDocument',
+        astMetadata,
         header: {
           kind: 'MapHeader',
           profile: {
@@ -1003,6 +1079,188 @@ describe('Check logic', () => {
         ],
         profileId: 'starwars/character-information',
         provider: mockProviderJson.name,
+      });
+    });
+  });
+
+  describe('when checking integration parameters', () => {
+    it('returns empty result if there are no integration parameters', async () => {
+      const mockSuperJson = new SuperJson({
+        profiles: {
+          [profile.name]: {
+            file: '',
+            providers: {
+              [provider]: {
+                file: '',
+              },
+            },
+          },
+        },
+        providers: {
+          [provider]: {
+            file: '',
+          },
+        },
+      });
+      expect(
+        checkIntegrationParameters(mockProviderJson, mockSuperJson)
+      ).toEqual({
+        issues: [],
+        kind: 'parameters',
+        provider: provider,
+      });
+    });
+
+    it('returns result with warnings if there are extra integration parameters in super.json', async () => {
+      const mockSuperJson = new SuperJson({
+        profiles: {
+          [profile.name]: {
+            file: '',
+            providers: {
+              [provider]: {
+                file: '',
+              },
+            },
+          },
+        },
+        providers: {
+          [provider]: {
+            file: '',
+            parameters: {
+              first: '$FIRST',
+              extra: 'hi, i am extra',
+            },
+          },
+        },
+      });
+
+      const mockProviderJson: ProviderJson = {
+        name: provider,
+        services: [{ id: 'test-service', baseUrl: 'service/base/url' }],
+        securitySchemes: [],
+        defaultService: 'test-service',
+        parameters: [
+          {
+            name: 'first',
+          },
+        ],
+      };
+      expect(
+        checkIntegrationParameters(mockProviderJson, mockSuperJson)
+      ).toEqual({
+        issues: [
+          {
+            kind: 'warn',
+            message: `Super.json defines parameter: extra which is not needed in provider ${provider}`,
+          },
+        ],
+        kind: 'parameters',
+        provider: provider,
+      });
+    });
+
+    it('returns result with errors if there are integration parameters missing in super.json', async () => {
+      const mockSuperJson = new SuperJson({
+        profiles: {
+          [profile.name]: {
+            file: '',
+            providers: {
+              [provider]: {
+                file: '',
+              },
+            },
+          },
+        },
+        providers: {
+          [provider]: {
+            file: '',
+            parameters: {
+              first: '$FIRST',
+              extra: 'hi, i am extra',
+            },
+          },
+        },
+      });
+
+      const mockProviderJson: ProviderJson = {
+        name: provider,
+        services: [{ id: 'test-service', baseUrl: 'service/base/url' }],
+        securitySchemes: [],
+        defaultService: 'test-service',
+        parameters: [
+          {
+            name: 'first',
+          },
+          {
+            name: 'second',
+          },
+        ],
+      };
+      expect(
+        checkIntegrationParameters(mockProviderJson, mockSuperJson)
+      ).toEqual({
+        issues: [
+          {
+            kind: 'warn',
+            message: `Super.json defines parameter: extra which is not needed in provider ${provider}`,
+          },
+          {
+            kind: 'error',
+            message: `Parameter second is not defined in super.json for provider ${provider}`,
+          },
+        ],
+        kind: 'parameters',
+        provider: provider,
+      });
+    });
+
+    it('returns result with error if there are not matching provider names super.json and provider.json', async () => {
+      const mockSuperJson = new SuperJson({
+        profiles: {
+          [profile.name]: {
+            file: '',
+            providers: {
+              [provider]: {
+                file: '',
+              },
+            },
+          },
+        },
+        providers: {
+          [provider]: {
+            file: '',
+            parameters: {
+              first: '$FIRST',
+            },
+          },
+        },
+      });
+
+      const mockProviderJson: ProviderJson = {
+        name: 'different',
+        services: [{ id: 'test-service', baseUrl: 'service/base/url' }],
+        securitySchemes: [],
+        defaultService: 'test-service',
+        parameters: [
+          {
+            name: 'first',
+          },
+          {
+            name: 'second',
+          },
+        ],
+      };
+      expect(
+        checkIntegrationParameters(mockProviderJson, mockSuperJson)
+      ).toEqual({
+        issues: [
+          {
+            kind: 'error',
+            message: `Provider different is not defined in super.json`,
+          },
+        ],
+        kind: 'parameters',
+        provider: 'different',
       });
     });
   });