+++ CHANGELOG.md
@@ -9,10 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Use provider.json from ast instead of sdk
 - Compile command now compiles every locally linked file in super.json
+- Compile command now clears cache before compilation
+- Check command checks for not matching provider name in provider.json and super.json
 
 ### Added
 - prepare integration parameters during configure command
 - install command fallbacks to parse profile source on profile AST validation fail
+- added `warn if update available` package
+- integration parameters setup check in `check` command and before publishing
 
 ## [0.0.26] - 2021-10-13
 ### Added